pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "optinist"
description = "Calcium Imaging Pipeline Tool"
version = "0.4.0"
readme = "README.md"
requires-python = "==3.8.*"
license = {file = "LICENSE"}
authors = [
  {name = "OIST"},
]
classifiers=[
  "Intended Audience :: Developers",
  "Intended Audience :: Education",
  "Intended Audience :: Science/Research",
  "Topic :: Scientific/Engineering",
  "Programming Language :: Python :: 3.7",
  "Programming Language :: Python :: 3.8",
  "Programming Language :: Python :: 3.9",
  "Operating System :: POSIX :: Linux",
  "Operating System :: Microsoft :: Windows",
  "Operating System :: MacOS :: MacOS X",
]
dependencies = [
  "fastapi==0.68.*",
  "uvicorn==0.15.*",
  "numpy==1.23.*",
  "aiofiles",
  "tqdm",
  "pyyaml",
  "imageio",
  "tifffile",
  "python-multipart",
  "future",
  "pynwb==2.2.0",
  "psutil",
  "nbstripout",
  "snakemake==7.24.*",
  "jinja2==3.0.*",
]

[project.optional-dependencies]
dev = [
  "black",
  "flake8",
  "isort",
  "build",
  "twine",
  "pre-commit",
]
doc = [
  "sphinx>=4.3.0,<4.6.0",
  "sphinxcontrib-apidoc",
  "sphinx_rtd_theme",
  "sphinx-prompt",
  "sphinx-pyproject",
  "sphinx-autodoc-typehints",
  "sphinx-copybutton==0.5.0",
  "sphinx-autobuild==2021.3.14",
  "myst-parser",
]
test = [
  "pytest",
  "pytest-env",
]

[project.urls]
Documentation = "https://optinist.readthedocs.io"
Repository = "https://github.com/oist/optinist"

[project.scripts]
run_optinist = "studio.__main__:main"

[tool.setuptools]
include-package-data = true
zip-safe = false
py-modules = ["Snakefile"]

[tool.setuptools.packages.find]
exclude = ["studio/tests*"]

[tool.setuptools.package-data]
frontend = [
  "build/*",
  "build/static/*",
  "build/static/css/*",
  "build/static/js/*",
  "build/static/media/*",
]
studio = [
  "services/nwb/nwb.yaml",
  "services/snakemake/snakemake.yaml",
  "wrappers/**/*.yaml",
  "Snakefile",
]

[tool.isort]
profile = "black"

[tool.pytest.ini_options]
pythonpath = "."
env = [
  "OPTINIST_DIR=studio/test_data",
]
